Implements an app file
The app, after initialised, receives a file as argument of a parse method that runs both LogParser and Printer to output the required specifications.

At this moment, the App is only a runner. The objective is to make it more user-friendly by requiring the user to only a few things, such as initialising the app and passing the file path.